+++ Craftsman/Builders/Endpoints/ControllerBuilder.cs
@@ -45,6 +45,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
 using System.Threading.Tasks;
 using System.Threading;
+using Asp.Versioning;
 using MediatR;
 
 [ApiController]