+++ Craftsman/Builders/ExtensionBuilders/ApiVersioningExtensionsBuilder.cs
@@ -23,25 +23,36 @@ public class ApiVersioningExtensionsBuilder
     {
         return @$"namespace {classNamespace};
 
-using FluentValidation.AspNetCore;
-using MediatR;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.OpenApi.Models;
-using System.Reflection;
+using Asp.Versioning;
+using Asp.Versioning.Conventions;
 
 public static class ApiVersioningServiceExtension
 {{
     public static void AddApiVersioningExtension(this IServiceCollection services)
     {{
-        services.AddApiVersioning(config =>
-        {{
-            // Default API Version
-            config.DefaultApiVersion = new ApiVersion(1, 0);
-            // use default version when version is not specified
-            config.AssumeDefaultVersionWhenUnspecified = true;
-            // Advertise the API versions supported for the particular endpoint
-            config.ReportApiVersions = true;
-        }});
+        services.AddApiVersioning(static options =>
+            {{
+                options.DefaultApiVersion = new ApiVersion(1, 0);
+                options.AssumeDefaultVersionWhenUnspecified = true;
+                options.ReportApiVersions = true;
+                options.ApiVersionReader = new UrlSegmentApiVersionReader();
+            }})
+            .AddMvc(static options =>
+            {{
+                options.Conventions.Add(new VersionByNamespaceConvention());
+            }})
+            .AddApiExplorer(
+                static options =>
+                {{
+                    // add the versioned api explorer, which also adds IApiVersionDescriptionProvider service
+                    // note: the specified format code will format the version as ""'v'major[.minor][-status]""
+                    options.GroupNameFormat = ""'v'VVV"";
+
+                    // note: this option is only necessary when versioning by url segment. the SubstitutionFormat
+                    // can also be used to control the format of the API version in route templates
+                    options.SubstituteApiVersionInUrl = true;
+                    options.AssumeDefaultVersionWhenUnspecified = true;
+                }});
     }}
 }}";
     }